extract rendering code into own module and slim down DataLayer further
