+++ app/assets/javascripts/oxalis/model/accessors/dataset_accessor.js
@@ -124,4 +124,8 @@ export function getBitDepth(layerInfo: DataLayerType): number {
   return parseInt(layerInfo.elementClass.substring(4), 10);
 }
 
+export function isSegmentation(dataset: APIDatasetType, layerName: string): boolean {
+  return getLayerByName(dataset, layerName).category === "segmentation";
+}
+
 export default {};
