+++ app/assets/javascripts/oxalis/model/bucket_data_handling/bucket_picker_strategies/flight_bucket_picker.js
@@ -7,17 +7,17 @@ import {
 import type { Vector3 } from "oxalis/constants";
 import type { Matrix4x4 } from "libs/mjs";
 import PriorityQueue from "js-priority-queue";
-import DataLayer from "oxalis/model/data_layer";
 import Utils from "libs/utils";
 import { M4x4, V3 } from "libs/mjs";
 import { getMatrixScale } from "oxalis/model/reducers/flycam_reducer";
 import constants from "oxalis/constants";
 import Store from "oxalis/store";
 import { getResolutions } from "oxalis/model/accessors/dataset_accessor";
+import type DataCube from "oxalis/model/bucket_data_handling/data_cube";
 import { getFallbackBuckets } from "./oblique_bucket_picker";
 
 export default function determineBucketsForFlight(
-  dataLayer: DataLayer,
+  cube: DataCube,
   bucketQueue: PriorityQueue,
   matrix: Matrix4x4,
   logZoomStep: number,
@@ -138,7 +138,7 @@ export default function determineBucketsForFlight(
   );
 
   for (const bucketAddress of traversedBuckets) {
-    const bucket = dataLayer.cube.getOrCreateBucket(bucketAddress);
+    const bucket = cube.getOrCreateBucket(bucketAddress);
 
     if (bucket.type !== "null") {
       const priority = V3.sub(bucketAddress, centerAddress).reduce((a, b) => a + Math.abs(b), 0);
